requirements.txt
pyspark
kafka-python
pandas
requests